+++ click_extended/decorators/check/__init__.py
@@ -1,5 +1,6 @@
 """Initialization file for the `click_extended.decorators.check` module."""
 
+from click_extended.decorators.check.conflicts import conflicts
 from click_extended.decorators.check.dependencies import dependencies
 from click_extended.decorators.check.ends_with import ends_with
 from click_extended.decorators.check.exclusive import exclusive
@@ -9,6 +10,7 @@ from click_extended.decorators.check.requires import requires
 from click_extended.decorators.check.starts_with import starts_with
 
 __all__ = [
+    "conflicts",
     "dependencies",
     "ends_with",
     "exclusive",